app/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
Automat==22.10.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
distro==1.8.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
openai==1.3.9
packaging==23.2
parsel==1.8.1
pathspec==0.12.1
platformdirs==4.1.0
Protego==0.3.0
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyDispatcher==2.0.7
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tldextract==5.1.1
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.1.0
w3lib==2.1.2
Werkzeug==3.0.1
yarl==1.9.4
zope.interface==6.1
